airflow_dags/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-docker
apache-airflow-providers-ssh
influxdb-client
joblib
matplotlib
numpy
openmeteo-requests
pandas
PyYAML
requests-cache
retry-requests
xgboost
scikit-learn
